feat: Use human-friendly timestamp for stale/ophan interva logging
